update utility test and restructure vuex store
