+++ src/util/comm.js
@@ -200,23 +200,6 @@ export function currEnv () {
   }
 }
 
-export function enableScroll () {
-  if (window.removeEventListener) {
-    window.removeEventListener('DOMMouseScroll', preventDefault, false)
-  }
-  window.onmousewheel = document.onmousewheel = null
-  window.onwheel = null
-  window.ontouchmove = null
-  document.onkeydown = null
-}
-function preventDefault (e) {
-  e = e || window.event
-  if (e.preventDefault) {
-    e.preventDefault()
-  }
-  e.returnValue = false
-}
-
 /*
  *  constructing and sending req to api to log client's behaviors through following functions:
  *    getClientOS()
@@ -435,12 +418,12 @@ export function removeClass (ele, cls) {
 //   return `<vpon vpon_ad_test="${mode}" vpon_ad_licensy_key="${key}" vpon_ad_format="${format}" debug="${debug}"></vpon>`
 // }
 
-export function updateCookie ({ currEnv }) {
+export function updateCookie () {
   return new Promise(resolve => {
     const cookie = Cookie.get('visited')    
-    if (currEnv === 'prod' && !cookie) {
-      // Cookie.set('visited', 'true', { expires: '3m' })
-    }
+    // if (currEnv === 'prod' && !cookie) {
+    //   Cookie.set('visited', 'true', { expires: '3m' })
+    // }
     resolve(cookie === 'true')
   })
 }
