+++ src/views/ArticleM.vue
@@ -384,11 +384,11 @@
               case 'MBCVR':
                 sendAdCoverGA('dfp')
                 if (adDisplayStatus === 'none') {
-                  updateCookie({ currEnv: currentInstance.dfpMode }).then((isVisited) => {
+                  updateCookie().then((isVisited) => {
                     currentInstance.showDfpCoverAd2Flag = !isVisited
                   })
                 } else {
-                  updateCookie({ currEnv: currentInstance.dfpMode }).then((isVisited) => {
+                  updateCookie().then((isVisited) => {
                     currentInstance.showDfpCoverAdFlag = !isVisited
                   })
                 }
