+++ src/views/Home.vue
@@ -271,11 +271,11 @@ export default {
             case 'LMBCVR':
               sendAdCoverGA('dfp')
               if (adDisplayStatus === 'none') {
-                updateCookie({ currEnv: currentInstance.dfpMode }).then((isVisited) => {
+                updateCookie().then((isVisited) => {
                   currentInstance.showDfpCoverAd2Flag = !isVisited
                 })
               } else {
-                updateCookie({ currEnv: currentInstance.dfpMode }).then((isVisited) => {
+                updateCookie().then((isVisited) => {
                   currentInstance.showDfpCoverAdFlag = !isVisited
                 })
               }
